Fixed packaging issue
